Add an alternative renderer for `bs stats` output
For scripting purposes a raw output is preferable over the rendered
table. This pull request introduces different renderer classes, the
existing `Table` renderer and the new `Raw` renderer. It displays
the tube data in a machine readable form (tab delimited rows).

The renderer can be selected by adding the `-e` option to the
`bs stats` command. If the option is missing the tool behaves as before.

All other options are unchanged so there're no compatibility issues.